+++ src/pushi/app/controllers/mail.py
@@ -1,0 +1,66 @@
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+
+# Hive Pushi System
+# Copyright (c) 2008-2024 Hive Solutions Lda.
+#
+# This file is part of Hive Pushi System.
+#
+# Hive Pushi System is free software: you can redistribute it and/or modify
+# it under the terms of the Apache License as published by the Apache
+# Foundation, either version 2.0 of the License, or (at your option) any
+# later version.
+#
+# Hive Pushi System is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# Apache License for more details.
+#
+# You should have received a copy of the Apache License along with
+# Hive Pushi System. If not, see <http://www.apache.org/licenses/>.
+
+__author__ = "João Magalhães <joamag@hive.pt>"
+""" The author(s) of the module """
+
+__copyright__ = "Copyright (c) 2008-2024 Hive Solutions Lda."
+""" The copyright for the module """
+
+__license__ = "Apache License, Version 2.0"
+""" The license for the module """
+
+import appier
+
+import pushi
+
+
+class MailController(appier.Controller):
+    @appier.private
+    @appier.route("/mails", "GET")
+    def list(self):
+        email = self.field("email", None)
+        event = self.field("event", None)
+        return self.state.mail_handler.subscriptions(email=email, event=event)
+
+    @appier.private
+    @appier.route("/mails", "POST")
+    def create(self):
+        auth = self.field("auth", None)
+        unsubscribe = self.field("unsubscribe", False, cast=bool)
+        mail = pushi.Mail.new()
+        mail = self.state.mail_handler.subscribe(
+            mail, auth=auth, unsubscribe=unsubscribe
+        )
+        return mail.map()
+
+    @appier.private
+    @appier.route("/mails/<email>", "DELETE")
+    def deletes(self, email):
+        mails = self.state.mail_handler.unsubscribes(email)
+        return dict(subscriptions=[mail.map() for mail in mails])
+
+    @appier.private
+    @appier.route(r"/mails/<email>/<regex('[\.\w-]+'):event>", "DELETE")
+    def delete(self, email, event):
+        force = self.field("force", False, cast=bool)
+        mail = self.state.mail_handler.unsubscribe(email, event=event, force=force)
+        return mail.map()